+++ lib/jtdetector.dart
@@ -1,14 +1,5 @@
 library jtdetector;
 
-import 'dart:ui' as ui show BoxHeightStyle, BoxWidthStyle;
-
-import 'package:flutter/cupertino.dart';
-import 'package:flutter/foundation.dart';
-import 'package:flutter/gestures.dart';
-import 'package:flutter/rendering.dart';
-import 'package:flutter/services.dart';
-import 'package:flutter/material.dart';
-
-part 'src/jHelpers.dart';
-part 'src/jTextDetector.dart';
-part 'src/jTextFieldDetector.dart';
+export 'src/jHelpers.dart';
+export 'src/jTextDetector.dart';
+export 'src/jTextFieldDetector.dart';